requirements.txt
python-telegram-bot==20.7
ollama==0.1.6
python-dotenv==1.0.0
aiohttp==3.9.1
aiosqlite==0.19.0
pytz==2024.1